Handle whitespace in parser, backwards compatibly
This allows whitespace between two hex chars (i.e., between two nibbles)
like `4 1` and between the digits of a quantifier like `{1 2}`. The
uselessness and stupidity of this feature is overwhelming. However, not
supporting it would break backwards compatibility. So let's drag it
along until a better reason for a major version increment presents
itself.